Add Makefile and config.h
The Makefile supports the usual DESTDIR and PREFIX to control the install
paths, as well as specific INCDIR and LIBDIR variables. The shared library
is built with a soname and minor version to match the upstream release.

Our config.h should provide sensible definitions to build and install on
either musl or glibc, and using either clang or gcc. It might need manual
adjustment on very old or non-Linux systems.